+++ src/bootctl/bootctl-cleanup.h
@@ -1,0 +1,6 @@
+/* SPDX-License-Identifier: LGPL-2.1-or-later */
+#pragma once
+
+#include "shared-forward.h"
+
+int verb_cleanup(int argc, char *argv[], void *userdata);
