+++ src/bootctl/meson.build
@@ -1,14 +1,17 @@
 # SPDX-License-Identifier: LGPL-2.1-or-later
 
 bootctl_sources = files(
+        'bootctl.c',
         'bootctl-install.c',
         'bootctl-random-seed.c',
         'bootctl-reboot-to-firmware.c',
         'bootctl-set-efivar.c',
         'bootctl-status.c',
         'bootctl-uki.c',
+        'bootctl-unlink.c',
+        'bootctl-cleanup.c',
         'bootctl-util.c',
-        'bootctl.c',
+        'bootspec-util.c',
 )
 
 executables += [
